changed data type to logical

+++ +mesh/+voxel/rasterize_fv.m
@@ -1,6 +1,6 @@
 function array = rasterize_fv( fv, points )
 
-array = double( VOXELISE( points{ 1 }, points{ 2 }, points{ 3 }, fv ) );
+array = VOXELISE( points{ 1 }, points{ 2 }, points{ 3 }, fv );
 
 end
